% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MetabinResult.R
\name{as.data.frame,MetabinResult-method}
\alias{as.data.frame,MetabinResult-method}
\title{Convert a \linkS4class{MetabinResult} to a \code{data.frame}}
\usage{
\S4method{as.data.frame}{MetabinResult}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{MetabinResult}.}

\item{row.names, optional, ...}{Unused; retained for S3 signature compatibility.}
}
\value{
A base \code{data.frame} of the assignments.
}
\description{
Preserves backwards compatibility with the \code{data.frame} return of
metabinR <= 1.x.
}
