% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabinR-package.R
\docType{package}
\name{metabinR-package}
\alias{metabinR}
\alias{metabinR-package}
\title{metabinR: Abundance and Compositional Based Binning of Metagenomes}
\description{
Provide functions for performing abundance and compositional based binning on metagenomic samples, directly from FASTA or FASTQ files. Functions are implemented in Java and called via rJava. Parallel implementation that operates directly on input FASTA/FASTQ files for fast execution. Inputs may be file paths or Biostrings/ShortRead sequence objects; results are returned as a MetabinResult S4 object wrapping cluster assignments, algorithm parameters, and input metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gkanogiannis/metabinR}
  \item Report bugs at \url{https://github.com/gkanogiannis/metabinR/issues}
}

}
\author{
\strong{Maintainer}: Anestis Gkanogiannis \email{anestis@gkanogiannis.com} (\href{https://orcid.org/0000-0002-6441-0688}{ORCID})

}
\keyword{internal}
