% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MetabinResult.R
\name{MetabinResult-accessors}
\alias{MetabinResult-accessors}
\alias{assignments}
\alias{nClusters}
\alias{parameters}
\alias{algorithm}
\alias{assignments,MetabinResult-method}
\alias{parameters,MetabinResult-method}
\alias{algorithm,MetabinResult-method}
\alias{nClusters,MetabinResult-method}
\title{Accessors for \linkS4class{MetabinResult}}
\usage{
assignments(x)

nClusters(x)

parameters(x)

algorithm(x)

\S4method{assignments}{MetabinResult}(x)

\S4method{parameters}{MetabinResult}(x)

\S4method{algorithm}{MetabinResult}(x)

\S4method{nClusters}{MetabinResult}(x)
}
\arguments{
\item{x}{A \linkS4class{MetabinResult} object.}
}
\value{
\code{assignments()} returns a \code{\link[S4Vectors]{DataFrame}}
  with the per-read cluster assignments and distances.
  \code{nClusters()} returns an integer scalar: the number of clusters
  inferred by the algorithm. \code{parameters()} returns the list of
  arguments passed to the binning function. \code{algorithm()} returns the
  algorithm tag (\code{"AB"}, \code{"CB"}, or \code{"ABxCB"}).
}
\description{
Accessors for \linkS4class{MetabinResult}
}
\examples{
res <- abundance_based_binning(
    system.file("extdata", "reads.metagenome.fasta.gz", package = "metabinR"),
    dryRun = TRUE, kMerSizeAB = 4, numOfClustersAB = 2
)
assignments(res)
nClusters(res)
algorithm(res)
parameters(res)$kMerSizeAB

}
