% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-MetabinResult.R
\docType{class}
\name{MetabinResult-class}
\alias{MetabinResult-class}
\alias{MetabinResult}
\alias{show,MetabinResult-method}
\title{MetabinResult: binning result container}
\usage{
\S4method{show}{MetabinResult}(object)
}
\arguments{
\item{object}{A \linkS4class{MetabinResult}.}
}
\value{
Objects of this class are returned by the binning functions.
  Use \code{\link{assignments}}, \code{\link{nClusters}},
  \code{\link{parameters}}, \code{\link{algorithm}}, or
  \code{as.data.frame()} to access the results.
}
\description{
An S4 class returned by [abundance_based_binning()],
[composition_based_binning()] and [hierarchical_binning()].
}
\section{Slots}{

\describe{
\item{\code{assignments}}{A \code{\link[S4Vectors]{DataFrame}} of cluster
assignments. The first column is \code{read_id}; subsequent columns
are algorithm-specific (see the corresponding binning function).}

\item{\code{parameters}}{Named list of the parameters passed to the algorithm.}

\item{\code{inputs}}{Character vector of input file paths that were processed.}

\item{\code{algorithm}}{Character scalar: one of \code{"AB"}, \code{"CB"},
\code{"ABxCB"}.}
}}

\examples{
res <- abundance_based_binning(
    system.file("extdata", "reads.metagenome.fasta.gz", package = "metabinR"),
    dryRun = TRUE, kMerSizeAB = 4, numOfClustersAB = 2
)
res
is(res, "MetabinResult")

}
